Don't track "builder" script.
We just ran into a case where `objcopy --debug-only` was causing
problems with our QEMU GDB. We removed it. But these changes aren't
there in the build scripts for old commits, which causes problems when
we test older revisions...

Need to think about this. Maybe build script should be in a separate
repository? For now, remove it from this repo to avoid conflicts.