Define cpuset helpers for swapon and swapoff.
Properly integrating them with the swapon code is more complicated, and
for another day. Even ignoring locks, adding a per-cgroup linked list
means we have a potential OOM when swapon-ing a device.

We've ignored that in this commit! It involves adding new error handling
code in swapon; we'll need a bit more understanding before we're ready
for that.