Add TODO for si swapin hotpath refcount issue.
How did the previous code work without using any ref counting there?
It dropped swap lock and si->lock...so it seems like there should be a
moment where si could be swapoff-ed.

Either there's a race condition in the current code, or we're missing
something and refcounting for nothing. Spurious hotpath refcounts are
bad, so...revisit this :)