Remove `si->users` percpu refcounts from cgroup accessors.
The users field is only used to prevent swapoff of a `swap_info_struct`.
It's not meant as a rigorous reference count; it's just a trigger for a
`struct completion`.

Do we have to worry about swapoff during a cgroup accessor?
If not, then we don't need _any_ refcounts there.