Refactor swaplist functionality for !CONFIG_CPUSETS case.
I'm reluctant to create a new file just to avoid sharing some
tiny methods...but it's either that or surround everything with a big
ifndef.

Also -- putting the decrement_subsequent_swap_prio function as non
static in the swaplist.c means that it'll be included even if it's not
needed (depends on CONFIG_CPUSETS). This needlessly increases binary
size on embedded architectures, so we might just duplicate those
functions? idk.